feat: move isStrapiSelectComponent to its own module
