+++ app/services/cms/components/StrapiSelect.ts
@@ -4,7 +4,6 @@ import {
   flattenStrapiErrors,
   StrapiErrorRelationSchema,
 } from "~/services/cms/flattenStrapiErrors";
-import { type StrapiFormComponent } from "~/services/cms/models/StrapiFormComponent";
 import { omitNull } from "~/util/omitNull";
 import { HasOptionalStrapiIdSchema } from "../models/HasStrapiId";
 import { StrapiSelectOptionSchema } from "../models/StrapiSelectOption";
@@ -24,12 +23,6 @@ type StrapiSelect = z.infer<typeof StrapiSelectSchema>;
 export const StrapiSelectComponentSchema = StrapiSelectSchema.extend({
   __component: z.literal("form-elements.select"),
 });
-
-type StrapiSelectComponent = z.infer<typeof StrapiSelectComponentSchema>;
-export const isStrapiSelectComponent = (
-  strapiContent: StrapiFormComponent,
-): strapiContent is StrapiSelectComponent =>
-  strapiContent.__component === "form-elements.select";
 
 /**
  * A Strapi "select" is really just a Radio Group
