+++ app/services/cms/models/isStrapiSelectComponent.ts
@@ -1,0 +1,10 @@
+import { type z } from "zod";
+import { type StrapiFormComponent } from "~/services/cms/models/StrapiFormComponent";
+import { type StrapiSelectComponentSchema } from "../components/StrapiSelect";
+
+type StrapiSelectComponent = z.infer<typeof StrapiSelectComponentSchema>;
+
+export const isStrapiSelectComponent = (
+  strapiContent: StrapiFormComponent,
+): strapiContent is StrapiSelectComponent =>
+  strapiContent.__component === "form-elements.select";
